Beginning of something big!
Added some basic files to handle the information and put up a structure on how it's going to work.

More information will be relealed tomorrow when I update readme file and make some comments inside the file.

Looking forward to this small little api project which should be finished in 2 days max! :+1: :book: