Friendlist, Player summaries, Userstats
Added interactions with the api handling with friendlist, player summaries and userstats for game.

Some module may need to major bug fixing.

Added a small module that's checking if it's a string and if its valid, it also has the ability to convert ints into string. I may publish this as an module if I find it very usefull!

Added some basic info on readme that tells I will be adding a better readme and docs later.

Made some more testing inside test.js

Imported all the modules that was added.